requirements.txt
# Project Scope Analyzer - Dependencies

# Core
anthropic>=0.18.0
python-dotenv>=1.0.0

# Data processing
pydantic>=2.5.0

# CLI and utilities
rich>=13.7.0
typer>=0.9.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0

# Document parsing (optional, for future enhancements)
python-docx>=1.1.0
pypdf>=3.17.0
markdown>=3.5.0

